rewritten entropy calculation
Hi,
your Entropy calculation only respects the choosen Action, but with this one every probability has an Influence on the Entropy, which means the Updates from the Entropy Loss are pushing into the generall more entropy direction and are not just maximizing the Entropy of the single choosen Action. I would really like to hear your thoughts on this!